Crate dynamixel_driver

Dynamixel Driver

Very simple driver for AX-12a Dynamixel motors as opposed to Dynamixel.rs this one supports sync writing and uses a simpler API But it doesn’t offer abstractions above different servos and protocols

Disclaimer

This software is not officially endorsed by ROBOTIS!

All product names, logos, and brands are property of their respective owners. All company, product and service names used in this website are for identification purposes only. Use of these names, logos, and brands does not imply endorsement.

Structs

DynamixelDriver

SyncCommand

SyncCommandFloat

Enums

DynamixelDriverError